Adding user exposed repo - Also improvements to user table, repo and service

+++ src/main/kotlin/com/p4ddy/paddycrm/application/user/UserApplicationService.kt
@@ -43,6 +43,17 @@ class UserApplicationService(
 	}
 
 	/**
+	 * Check if the password is valid for the user with the given id
+	 *
+	 * @param userId The id of the user to check the password against
+	 * @param password The password to check for validity
+	 * @return If the password is valid
+	 */
+	fun checkIfPasswordIsValid(userId: Int, password: String): Boolean {
+		return userRepo.checkPasswordValid(userId, password)
+	}
+
+	/**
 	 * Create a new sales rep user and save it to the database
 	 *
 	 * @param lastName The last name of the user
@@ -59,6 +70,13 @@ class UserApplicationService(
 		email: String,
 		managerId: Int
 	): User {
+		if (UserSingleton.user == null) {
+			throw Exception("Can't create user when no user is logged in")
+		}
+		if (UserSingleton.user!!.userType != UserTypes.ADMIN) {
+			throw Exception("Only system administrators may create users")
+		}
+
 		val user = User(lastName, firstName, password, email, UserTypes.SALESREP, managerId)
 		return userRepo.save(user)
 	}
@@ -79,6 +97,13 @@ class UserApplicationService(
 		email: String,
 		managerId: Int
 	): User {
+		if (UserSingleton.user == null) {
+			throw Exception("Can't create user when no user is logged in")
+		}
+		if (UserSingleton.user!!.userType != UserTypes.ADMIN) {
+			throw Exception("Only system administrators may create users")
+		}
+
 		val user = User(lastName, firstName, password, email, UserTypes.MANAGER)
 		return userRepo.save(user)
 	}
@@ -99,6 +124,16 @@ class UserApplicationService(
 		email: String,
 		managerId: Int
 	): User {
+		// So we can initially create an admin account upon first start without getting an Exception
+		if (userRepo.findAll().isNotEmpty()) {
+			if (UserSingleton.user == null) {
+				throw Exception("Can't create user when no user is logged in")
+			}
+			if (UserSingleton.user!!.userType != UserTypes.ADMIN) {
+				throw Exception("Only system administrators may create users")
+			}
+		}
+
 		val user = User(lastName, firstName, password, email, UserTypes.ADMIN)
 		return userRepo.save(user)
 	}
@@ -110,6 +145,13 @@ class UserApplicationService(
 	 * @return The updated user object
 	 */
 	fun updateUser(user: User): User {
+		if (UserSingleton.user == null) {
+			throw Exception("Can't update a user when no user is currently logged in")
+		}
+		if (UserSingleton.user!!.userType != UserTypes.ADMIN && UserSingleton.user!!.userId != user.userId) {
+			throw Exception("Only admins and the users themselves may update a user record")
+		}
+
 		return userRepo.update(user)
 	}
 
@@ -119,6 +161,13 @@ class UserApplicationService(
 	 * @param user: The user to delete
 	 */
 	fun deleteUser(user: User) {
+		if (UserSingleton.user == null) {
+			throw Exception("Can't update a user when no user is currently logged in")
+		}
+		if (UserSingleton.user!!.userType != UserTypes.ADMIN) {
+			throw Exception("Only admins may delete a user record")
+		}
+
 		return userRepo.delete(user)
 	}
 }
